Consolidate pool interface into 3_0
