+++ src/dbus_api/pool/mod.rs
@@ -15,11 +15,7 @@ use crate::{
 };
 
 mod fetch_properties_3_0;
-mod pool_2_0;
-mod pool_2_1;
-mod pool_2_3;
-mod pool_2_4;
-mod pool_2_5;
+mod pool_3_0;
 mod shared;
 
 pub fn create_dbus_pool<'a>(
@@ -40,79 +36,23 @@ pub fn create_dbus_pool<'a>(
         )
         .introspectable()
         .add(
-            f.interface(consts::POOL_INTERFACE_NAME, ())
-                .add_m(pool_2_0::create_filesystems_method(&f))
-                .add_m(pool_2_0::destroy_filesystems_method(&f))
-                .add_m(pool_2_0::snapshot_filesystem_method(&f))
-                .add_m(pool_2_0::add_blockdevs_method(&f))
-                .add_m(pool_2_0::add_cachedevs_method(&f))
-                .add_m(pool_2_0::rename_method(&f))
-                .add_p(pool_2_0::name_property(&f))
-                .add_p(pool_2_0::uuid_property(&f)),
-        )
-        .add(
-            f.interface(consts::POOL_INTERFACE_NAME_2_1, ())
-                .add_m(pool_2_0::create_filesystems_method(&f))
-                .add_m(pool_2_0::destroy_filesystems_method(&f))
-                .add_m(pool_2_0::snapshot_filesystem_method(&f))
-                .add_m(pool_2_0::add_blockdevs_method(&f))
-                .add_m(pool_2_1::init_cache_method(&f))
-                .add_m(pool_2_1::add_cachedevs_method(&f))
-                .add_m(pool_2_0::rename_method(&f))
-                .add_p(pool_2_0::name_property(&f))
-                .add_p(pool_2_0::uuid_property(&f))
-                .add_p(pool_2_1::encrypted_property(&f)),
-        )
-        .add(
-            f.interface(consts::POOL_INTERFACE_NAME_2_3, ())
-                .add_m(pool_2_0::create_filesystems_method(&f))
-                .add_m(pool_2_0::destroy_filesystems_method(&f))
-                .add_m(pool_2_0::snapshot_filesystem_method(&f))
-                .add_m(pool_2_0::add_blockdevs_method(&f))
-                .add_m(pool_2_3::bind_clevis_method(&f))
-                .add_m(pool_2_3::unbind_clevis_method(&f))
-                .add_m(pool_2_1::init_cache_method(&f))
-                .add_m(pool_2_1::add_cachedevs_method(&f))
-                .add_m(pool_2_0::rename_method(&f))
-                .add_p(pool_2_0::name_property(&f))
-                .add_p(pool_2_0::uuid_property(&f))
-                .add_p(pool_2_1::encrypted_property(&f)),
-        )
-        .add(
-            f.interface(consts::POOL_INTERFACE_NAME_2_4, ())
-                .add_m(pool_2_0::create_filesystems_method(&f))
-                .add_m(pool_2_0::destroy_filesystems_method(&f))
-                .add_m(pool_2_0::snapshot_filesystem_method(&f))
-                .add_m(pool_2_0::add_blockdevs_method(&f))
-                .add_m(pool_2_3::bind_clevis_method(&f))
-                .add_m(pool_2_3::unbind_clevis_method(&f))
-                .add_m(pool_2_1::init_cache_method(&f))
-                .add_m(pool_2_1::add_cachedevs_method(&f))
-                .add_m(pool_2_4::bind_keyring_method(&f))
-                .add_m(pool_2_4::unbind_keyring_method(&f))
-                .add_m(pool_2_0::rename_method(&f))
-                .add_p(pool_2_0::name_property(&f))
-                .add_p(pool_2_0::uuid_property(&f))
-                .add_p(pool_2_1::encrypted_property(&f)),
-        )
-        .add(
             f.interface(consts::POOL_INTERFACE_NAME_3_0, ())
-                .add_m(pool_2_0::create_filesystems_method(&f))
-                .add_m(pool_2_0::destroy_filesystems_method(&f))
-                .add_m(pool_2_0::snapshot_filesystem_method(&f))
-                .add_m(pool_2_0::add_blockdevs_method(&f))
-                .add_m(pool_2_3::bind_clevis_method(&f))
-                .add_m(pool_2_3::unbind_clevis_method(&f))
-                .add_m(pool_2_1::init_cache_method(&f))
-                .add_m(pool_2_1::add_cachedevs_method(&f))
-                .add_m(pool_2_4::bind_keyring_method(&f))
-                .add_m(pool_2_4::unbind_keyring_method(&f))
-                .add_m(pool_2_5::rebind_keyring_method(&f))
-                .add_m(pool_2_5::rebind_clevis_method(&f))
-                .add_m(pool_2_0::rename_method(&f))
-                .add_p(pool_2_0::name_property(&f))
-                .add_p(pool_2_0::uuid_property(&f))
-                .add_p(pool_2_1::encrypted_property(&f)),
+                .add_m(pool_3_0::create_filesystems_method(&f))
+                .add_m(pool_3_0::destroy_filesystems_method(&f))
+                .add_m(pool_3_0::snapshot_filesystem_method(&f))
+                .add_m(pool_3_0::add_blockdevs_method(&f))
+                .add_m(pool_3_0::bind_clevis_method(&f))
+                .add_m(pool_3_0::unbind_clevis_method(&f))
+                .add_m(pool_3_0::init_cache_method(&f))
+                .add_m(pool_3_0::add_cachedevs_method(&f))
+                .add_m(pool_3_0::bind_keyring_method(&f))
+                .add_m(pool_3_0::unbind_keyring_method(&f))
+                .add_m(pool_3_0::rebind_keyring_method(&f))
+                .add_m(pool_3_0::rebind_clevis_method(&f))
+                .add_m(pool_3_0::rename_method(&f))
+                .add_p(pool_3_0::name_property(&f))
+                .add_p(pool_3_0::uuid_property(&f))
+                .add_p(pool_3_0::encrypted_property(&f)),
         )
         .add(
             f.interface(consts::PROPERTY_FETCH_INTERFACE_NAME_3_0, ())
@@ -133,25 +73,6 @@ pub fn get_pool_properties(
     pool: &dyn Pool,
 ) -> InterfacesAddedThreadSafe {
     initial_properties! {
-        consts::POOL_INTERFACE_NAME => {
-            consts::POOL_NAME_PROP => shared::pool_name_prop(pool_name),
-            consts::POOL_UUID_PROP => uuid_to_string!(pool_uuid)
-        },
-        consts::POOL_INTERFACE_NAME_2_1 => {
-            consts::POOL_NAME_PROP => shared::pool_name_prop(pool_name),
-            consts::POOL_UUID_PROP => uuid_to_string!(pool_uuid),
-            consts::POOL_ENCRYPTED_PROP => shared::pool_enc_prop(pool)
-        },
-        consts::POOL_INTERFACE_NAME_2_3 => {
-            consts::POOL_NAME_PROP => shared::pool_name_prop(pool_name),
-            consts::POOL_UUID_PROP => uuid_to_string!(pool_uuid),
-            consts::POOL_ENCRYPTED_PROP => shared::pool_enc_prop(pool)
-        },
-        consts::POOL_INTERFACE_NAME_2_4 => {
-            consts::POOL_NAME_PROP => shared::pool_name_prop(pool_name),
-            consts::POOL_UUID_PROP => uuid_to_string!(pool_uuid),
-            consts::POOL_ENCRYPTED_PROP => shared::pool_enc_prop(pool)
-        },
         consts::POOL_INTERFACE_NAME_3_0 => {
             consts::POOL_NAME_PROP => shared::pool_name_prop(pool_name),
             consts::POOL_UUID_PROP => uuid_to_string!(pool_uuid),
