--- src/dbus_api/pool/pool_2_0/mod.rs
@@ -1,9 +1,0 @@
-mod api;
-mod methods;
-mod props;
-
-pub use api::{
-    add_blockdevs_method, add_cachedevs_method, create_filesystems_method,
-    destroy_filesystems_method, name_property, rename_method, snapshot_filesystem_method,
-    uuid_property,
-};
