--- src/dbus_api/pool/pool_2_1/mod.rs
@@ -1,5 +1,0 @@
-mod api;
-mod methods;
-mod props;
-
-pub use api::{add_cachedevs_method, encrypted_property, init_cache_method};
