--- src/dbus_api/pool/pool_2_3/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{bind_clevis_method, unbind_clevis_method};
