--- src/dbus_api/pool/pool_2_4/api.rs
@@ -1,31 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-use dbus_tree::{Factory, MTSync, Method};
-
-use crate::dbus_api::{
-    pool::pool_2_4::methods::{bind_keyring, unbind_keyring},
-    types::TData,
-};
-
-pub fn bind_keyring_method(f: &Factory<MTSync<TData>, TData>) -> Method<MTSync<TData>, TData> {
-    f.method("BindKeyring", (), bind_keyring)
-        .in_arg(("key_desc", "s"))
-        // b: Indicates if new keyring bindings were added
-        //
-        // Rust representation: bool
-        .out_arg(("results", "b"))
-        .out_arg(("return_code", "q"))
-        .out_arg(("return_string", "s"))
-}
-
-pub fn unbind_keyring_method(f: &Factory<MTSync<TData>, TData>) -> Method<MTSync<TData>, TData> {
-    f.method("UnbindKeyring", (), unbind_keyring)
-        // b: Indicates if keyring bindings were removed
-        //
-        // Rust representation: bool
-        .out_arg(("results", "b"))
-        .out_arg(("return_code", "q"))
-        .out_arg(("return_string", "s"))
-}
