--- src/dbus_api/pool/pool_2_4/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{bind_keyring_method, unbind_keyring_method};
