--- src/dbus_api/pool/pool_2_5/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{rebind_clevis_method, rebind_keyring_method};
