+++ src/dbus_api/pool/pool_3_0/mod.rs
@@ -1,0 +1,10 @@
+mod api;
+mod methods;
+mod props;
+
+pub use api::{
+    add_blockdevs_method, add_cachedevs_method, bind_clevis_method, bind_keyring_method,
+    create_filesystems_method, destroy_filesystems_method, encrypted_property, init_cache_method,
+    name_property, rebind_clevis_method, rebind_keyring_method, rename_method,
+    snapshot_filesystem_method, unbind_clevis_method, unbind_keyring_method, uuid_property,
+};
